requirements.txt
qdrant-client
openai
flask
python-dotenv
sarvamai
pymongo